moved registry tests to correct folders
