+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Form/tests/registries/FieldRegistry.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable flowtype/require-valid-file-annotation */
 import React from 'react';
-import fieldRegistry from '../registries/FieldRegistry';
+import fieldRegistry from '../../registries/FieldRegistry';
 
 beforeEach(() => {
     fieldRegistry.clear();
