+++ src/Sulu/Bundle/AdminBundle/Resources/js/services/Router/tests/registries/RouteRegistry.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable flowtype/require-valid-file-annotation */
-import routeRegistry from '../registries/RouteRegistry';
+import routeRegistry from '../../registries/RouteRegistry';
 
 beforeEach(() => {
     routeRegistry.clear();
